index context file created for better imports, login utils created
