+++ src/components/search/Search.jsx
@@ -1,4 +1,4 @@
-import { useFilterProducts } from "../../contexts/filter-context/filter-context";
+import { useFilterProducts } from "../../contexts/index-context";
 
 export default function Search() {
   const { state, dispatch } = useFilterProducts();
