+++ src/contexts/index-context.jsx
@@ -1,0 +1,15 @@
+import { useCart, CartProvider } from "./cart-context/cart-context";
+import { useLogin, LoginProvider } from "./login-context/login-context";
+import {
+  useFilterProducts,
+  FilterProductsProvider,
+} from "./filter-context/filter-context";
+
+export {
+  useCart,
+  CartProvider,
+  useLogin,
+  LoginProvider,
+  useFilterProducts,
+  FilterProductsProvider,
+};
