+++ src/index.js
@@ -4,9 +4,7 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
-import { FilterProductsProvider } from "./contexts/filter-context/filter-context";
-import { LoginProvider } from "./contexts/login-context/login-context";
-import { CartProvider } from "./contexts/cart-context/cart-context";
+import { CartProvider,LoginProvider,FilterProductsProvider } from "./contexts/index-context";
 
 // Call make Server
 makeServer();
